add mpi

+++ tasks/seq/gnitienko_k_bellman-ford_algorithm/src/ops_seq.cpp
@@ -37,7 +37,7 @@ bool gnitienko_k_bellman_ford_algorithm_seq::BellmanFordAlgSeq::validation() {
 bool gnitienko_k_bellman_ford_algorithm_seq::BellmanFordAlgSeq::Iteration(std::vector<int>& paths) {
   bool changed = false;
   for (size_t i = 0; i < V; ++i) {
-    for (size_t j = row_ptr[i]; j < row_ptr[i + 1]; ++j) {
+    for (int j = row_ptr[i]; j < row_ptr[i + 1]; ++j) {
       int v = columns[j];
       int weight = values[j];
       if (paths[i] != INF && paths[i] + weight < paths[v]) {
@@ -52,7 +52,7 @@ bool gnitienko_k_bellman_ford_algorithm_seq::BellmanFordAlgSeq::Iteration(std::v
 
 bool gnitienko_k_bellman_ford_algorithm_seq::BellmanFordAlgSeq::check_negative_cycle() {
   for (size_t i = 0; i < V; ++i) {
-    for (size_t j = row_ptr[i]; j < row_ptr[i + 1]; ++j) {
+    for (int j = row_ptr[i]; j < row_ptr[i + 1]; ++j) {
       int v = columns[j];
       int weight = values[j];
       if (shortest_paths[i] != INF && shortest_paths[i] + weight < shortest_paths[v]) {
@@ -68,7 +68,7 @@ bool gnitienko_k_bellman_ford_algorithm_seq::BellmanFordAlgSeq::run() {
   internal_order_test();
 
   bool changed = false;
-  for (int i = 0; i < V - 1; i++) {
+  for (size_t i = 0; i < V - 1; i++) {
     changed = Iteration(shortest_paths);
     if (!changed) break;
   }
@@ -77,7 +77,7 @@ bool gnitienko_k_bellman_ford_algorithm_seq::BellmanFordAlgSeq::run() {
     return false;
   }
 
-  for (int i = 0; i < V; i++) {
+  for (size_t i = 0; i < V; i++) {
     if (shortest_paths[i] == INF) shortest_paths[i] = 0;
   }
   return true;
@@ -85,86 +85,8 @@ bool gnitienko_k_bellman_ford_algorithm_seq::BellmanFordAlgSeq::run() {
 
 bool gnitienko_k_bellman_ford_algorithm_seq::BellmanFordAlgSeq::post_processing() {
   internal_order_test();
-  for (int i = 0; i < V; i++) {
+  for (size_t i = 0; i < V; i++) {
     reinterpret_cast<int*>(taskData->outputs[0])[i] = shortest_paths[i];
   }
   return true;
 }
-
-//bool gnitienko_k_bellman_ford_algorithm_seq::BellmanFordAlgSeq::pre_processing() {
-//  internal_order_test();
-//  auto* input_matrix = reinterpret_cast<int*>(taskData->inputs[0]);
-//  V = taskData->inputs_count[0];
-//  E = taskData->inputs_count[1];
-//  adjacency_matrix.resize(V, std::vector<int>(V));
-//  for (int i = 0; i < V; i++) {
-//    for (int j = 0; j < V; j++) {
-//      adjacency_matrix[i][j] = input_matrix[i * V + j];
-//    }
-//  }
-//  shortest_paths.resize(V, (INT_MAX - 1)/2);
-//  shortest_paths[0] = 0;
-//  return true;
-//}
-//
-//bool gnitienko_k_bellman_ford_algorithm_seq::BellmanFordAlgSeq::validation() {
-//  internal_order_test();
-//  return taskData->inputs_count[0] < taskData->inputs_count[1] &&
-//         taskData->inputs_count[0] == taskData->outputs_count[0];
-//}
-//
-//bool gnitienko_k_bellman_ford_algorithm_seq::BellmanFordAlgSeq::Iteration(std::vector<int>& paths) {
-//  bool changed = false;
-//  for (int i = 0; i < V; i++) {
-//    for (int j = 0; j < V; j++) {
-//      if (i != j) {
-//        if (adjacency_matrix[i][j] != 0 && adjacency_matrix[i][j] + paths[i] < paths[j]) {
-//          paths[j] = adjacency_matrix[i][j] + paths[i];
-//          changed = true;
-//        }
-//      }
-//    }
-//  }
-//
-//  return changed;
-//}
-//
-//bool gnitienko_k_bellman_ford_algorithm_seq::BellmanFordAlgSeq::check_negative_cycle() {
-//  for (int i = 0; i < V; i++) {
-//    for (int j = 0; j < V; j++) {
-//      if (i != j && adjacency_matrix[i][j] != 0) {
-//        if (shortest_paths[i] + adjacency_matrix[i][j] < shortest_paths[j]) {
-//          std::cerr << "Negative cycle detected!" << std::endl;
-//          return true;
-//        }
-//      }
-//    }
-//  }
-//  return false;
-//}
-//
-//bool gnitienko_k_bellman_ford_algorithm_seq::BellmanFordAlgSeq::run() {
-//  internal_order_test();
-//  bool stop = false;
-//
-//  for (int i = 0; i < V - 1; i++) {
-//    stop = Iteration(shortest_paths);
-//    if (!stop) break;
-//  }
-//
-//  if (check_negative_cycle()) {
-//    return false;
-//  }
-//  for (int i = 0; i < V; i++)
-//    if (shortest_paths[i] == (INT_MAX - 1) / 2) shortest_paths[i] = NULL;
-//
-//  return true;
-//}
-//
-//bool gnitienko_k_bellman_ford_algorithm_seq::BellmanFordAlgSeq::post_processing() {
-//  internal_order_test();
-//  for (int i = 0; i < V; i++) {
-//    reinterpret_cast<int*>(taskData->outputs[0])[i] = shortest_paths[i];
-//  }
-//  return true;
-//}
